move and rename searchContext to SearchProvider
