+++ providers/SearchProvider.js
